+++ src/fields/Products.php
@@ -9,7 +9,12 @@ namespace craft\shopify\fields;
 
 use Craft;
 use craft\fields\BaseRelationField;
+use craft\helpers\Gql;
 use craft\shopify\elements\Product;
+use craft\shopify\gql\arguments\elements\Product as ProductArguments;
+use craft\shopify\gql\interfaces\elements\Product as ProductInterface;
+use craft\shopify\gql\resolvers\elements\Product as ProductResolver;
+use GraphQL\Type\Definition\Type;
 
 /**
  * Class Shopify Product Field
@@ -43,5 +48,20 @@ class Products extends BaseRelationField
     public static function elementType(): string
     {
         return Product::class;
+    }
+
+    /**
+     * @inheritdoc
+     * @since 7.1.0
+     */
+    public function getContentGqlType(): Type|array
+    {
+        return [
+            'name' => $this->handle,
+            'type' => Type::listOf(ProductInterface::getType()),
+            'args' => ProductArguments::getArguments(),
+            'resolve' => ProductResolver::class . '::resolve',
+            'complexity' => Gql::eagerLoadComplexity(),
+        ];
     }
 }